Add files via upload
Script for automatically touching files on hopper to avoid them being deleted.  In header, need to adjust output name and email address.  Also need to input proper directory names to the "find" command.  After "touching" files, the script runs "expiredfiles_recheck.sh" which double checks that the files are no longer scheduled for deletion.  "expiredfiles_recheck.sh" may take a while (up to a couple of hours depending on number of filed) and you can choose to remove it from the script